memsup: Return correct amount of total system memory on macOS
On macOS, `memsup:get_system_memory_data/0` and `memsup:get_memory_data/0` returns
a much too low number for the amount system memory.

The bug is in a function that tried to determine the amount of memory by summing
numbers from the output of the `vm_stat` command.

Instead of trying to fix the summation, we will change approach and use the port
program to retrieve the information. Using the Mach APIs the total amount of memory
can be retrieved directly without any summing.

https://bugs.erlang.org/browse/ERL-1327