MB-45793 io_lib:format/3 with chars_limit opt on some terms falls into a loop
The bug is in the function io_lib_pretty:intermediate/6 which is called by
io_lib:format/3.

The algorithm works in intermediate/6 works as follow:
1) Start with an initial Depth D0 = 1. Number of chars used to print the term (L=0).
2) Walk the term, and update L.
3) If L < chars_limit, update D0 = 2*D0. And repeat from step 1.
4) If L > chars_limit, returned the trimmed term.

Due to a bug, L is always less than chars_limit and we end up in a loop.

L in step 2 is computed by io_lib_pretty:find_upper, using an auxiliary
length T. The bug is in the computation of T, specifically for nest List
terms like, [1000,[999,[998 ...]]].