stdlib: Fix a loop when using io_lib option 'chars_limit'
See GH-4824: https://github.com/erlang/otp/pull/4824.

The fix of ERL-967, see https://github.com/erlang/otp/issues/4307,
proved to be inaccurate.

Thanks to Hareen (kvsrh) for tracking down the bug. The fix in
https://github.com/erlang/otp/pull/4842 is included.

A clause that breaks the infinite loop is also added. It will hide
further bugs, but it is deemed more important not to loop.