Provide creation value when undefined for otp23
During handshake, the 'creation' value is used for otp23 and is usually
undefined for some reason. The creation value seems to come from a NIF
and it's somewhat unknown as to why this happens. The old handshake
did not use the creation value at all, so this was not an issue. Have
confirmed that we are providing the creation value in our code, but it
is coming from an internal function instead. The BIF
erts_internal:get_creation/0 is also new in erlang 23.