MB-48998: Request client cert passed improper list of ciphers
Tlsv1.3 introduced a different standard regarding client certificate
requests during the server hello response. In the old format a
structure named hash_sign_algos was used and eventually passed to the
serializer. In tlsv1.3 that structure changed to signature_algorithms
where it also included signature schemes. This issue caused invalid
tls packets to be sent during the handshake process for a tlsv1.2
connection. This issue only arises if the server is used to serve
tlsv1.2 and tlsv1.3 connections. The difference being that the extra
information provided for tlsv1.3 is never removed for tlsv1.2 connections.